requirements.txt
diffusers==0.11.1
transformers
scipy 
ftfy 
accelerate
pillow